P4-4017: add easier "Get PM" button/link to /msg/inbox.
