+++ engage/msgs/views/inbox.py
@@ -1,13 +1,9 @@
 import logging
 
-from django.conf import settings
-from django.contrib import messages
-from django.http import HttpResponseRedirect
 from django.urls import reverse
-from django.utils.safestring import mark_safe
-from django.utils.translation import gettext_lazy as _
 
 from engage.utils.class_overrides import ClassOverrideMixinMustBeFirst
+from engage.utils.strings import sanitize_text
 
 from temba.msgs.views import InboxView
 
@@ -15,22 +11,59 @@ from temba.msgs.views import InboxView
 logger = logging.getLogger(__name__)
 
 class MsgInboxViewOverrides(ClassOverrideMixinMustBeFirst, InboxView):
+    """
+    Sanitize the list of message contents to remove zero-width spaces and such.
+    """
+
+    """
+    def pre_process(self, request, *args, **kwargs):
+        self.getOrigClsAttr('pre_process')(self, request, *args, **kwargs)
+        # give us the ability to override the pagination (super helpful in debugging)
+        if 'r' in self.request.GET:
+            self.refresh = self.request.GET['r']
+        if 'nor' in self.request.GET:
+            self.refresh = None
+        if 'page_by' in self.request.GET:
+            self.paginate_by = self.request.GET['page_by']
+    """
+
+    def _sanitizeMsgList(self, aList):
+        """
+        Ensure HTML in messages do not bork our display.
+        :param aList: the list of message to process.
+        :return: the processed list.
+        """
+        for theMsg in aList:
+            if theMsg and hasattr(theMsg, 'text') and theMsg.text is not None:
+                if isinstance(theMsg.text, str):
+                    theMsg.text = sanitize_text(theMsg.text)
+                elif isinstance(theMsg.text, dict):
+                    for key, val in theMsg.text.items():
+                        if isinstance(val, str):
+                            theMsg.text[key] = sanitize_text(val)
+                        #fi
+                    #rof
+                #fi
+            #fi
+        #rof
+        return aList
+    #enddef _sanitizeMsgList
+
+    def get_context_data(self, **kwargs):
+        context = self.getOrigClsAttr('get_context_data')(self, **kwargs)
+        context['object_list'] = self._sanitizeMsgList(context['object_list'])
+        return context
+    #enddef get_context_data
 
     def get_gear_links(self):
         links = self.getOrigClsAttr('get_gear_links')(self)
-
-        # append the Purge Outbox link as a button
         links.append(
             dict(
-                id="action-jump2pm",
                 title="Get PM",
                 as_btn=True,
-                #href='/channels/types/postmaster/claim',
                 href=reverse("channels.types.postmaster.claim"),
-                js_class="button-info",
             )
         )
-
         return links
     #enddef get_gear_links
